requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Database
PyMySQL==1.1.0
cryptography==41.0.7  # For Fernet encryption + PyMySQL SSL

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0

# Password hashing
Werkzeug==3.0.1

# Environment variables
python-dotenv==1.0.0

# Production server
gunicorn==21.2.0

# Development/testing
pytest==7.4.3
pytest-flask==1.3.0
